+++ NEWS.md
@@ -1,3 +1,17 @@
+# common 1.0.3
+
+* Added `file.find()` function to look for files on the file system.
+* Added `dir.find()` function to look for directories on the file system.
+* Added `find.names()` function to look for variable names on a data frame.
+* Added `copy.attributes()` function to copy column attributes from 
+one data frame to another.
+* Added `supsc()` and `subsc()` functions to get UTF-8 superscript and subscript
+codes from normal characters.
+* Fixed bug on `v()` function when first item had more than one part.
+* Fixed bug on `sort.data.frame()` that was causing it to lose column attributes.
+* Restructured test and code files.
+* Improved documentation.
+
 # common 1.0.1
 
 * Published to CRAN.
@@ -5,7 +19,7 @@
 # common 0.0.9001
 
 * Add news.
-* Add readment.
+* Add readme.
 * Complete documentation.
 * Build pkgdown site.
 
@@ -15,7 +29,7 @@
 - `labels.data.frame()`: Adds, views and removes labels on a data frame.
 - `sort.data.frame()`: Sorts a data frame.
 - `%p%`: An infix operator for the `paste0()` function.
-- `%eq`: An infix operator to determine equality between two objects.
+- `%eq%`: An infix operator to determine equality between two objects.
 - `roundup()`: A function that rounds 5 away from zero.
 - `Sys.path()`: A function to get the path of the currently executing program.
 - `v()`: A quoting function for non-standard evaluation.